[Spring 지하철 노선도 - 3단계] 잉(유영우) 미션 제출합니다. (#304)
* docs: 3단계 기능 목록 작성

* feat: SectionDao 생성(저장, 삭제 기능)

* feat: SectionDao lineId에 따른 조회기능 추가

* feat: StationDao findById() 기능 추가

* feat: StationService id에 따른 지하철 역 단일 조회 기능 추가

* feat: Sections 생성 및 List<Section> -> List<Long>으로 변환하는 기능 구현
StationId로 지하철역 이름 알아내기 위해

* refactor: List<Section> -> Sections로 return 값 변경

* feat: SectionService 생성
(구간등록 및 라인별 지허철역들 조회 기능 추가)

* refactor: .sql 파일 변경

* feat: 스프링 어노테이션 추가

* fix: Line과 1:N 관계 설정에 따른 에러 수정

* feat: response에 관련 노선 지하철역들도 내보내주기

* feat: 상행, 하행 역 기준으로 조회하는 기능, 상행 역 기준으로 삭제하는 기능 추가

* feat: 지하철 호선과 상행역을 기준으로 구간의 하행역을 바꾸는 기능

* feat: 지하철 호선과 지하철역에 따라 구간을 조회하는 기능

* feat: SectionService 구간삭제 기능

* refactor: SectionService createSection() 매개변수 변경

* feat: db에 상행, 하행역 있는지 확인하는 기능 추가 및 적용

* feat: db에 상행역 있는지 확인하는 기능 과 db에 하행역 있는지 확인하는 기능

* feat: 해당 라인과 지하철역 2개에 관련된 구간들을 조회하는 기능

* feat: 지하철 구간들(sections) 중에서 요청 상행, 하행역과 연관될 구간(section)을 정하는 기능

* feat: 지하철 호선과 행역을 기준으로 구간의 상행역을 수정하는 기능

* style: 바뀐 메서드명 적용

* feat: 구간 등록 기능 서비스에 추가

* refactor: 의미상 중복되는 쿼리 제거

* fix: 접근제어자 적절히 변경

* feat: SectionController 구현

* style: 불필요한 주석 제거

* docs: 3단계 기능목록 체크

* test: test 설명위해 이름 지정해주기

* refactor: 함수 명 오타 수정

* refactor(SectionService): 불필요한 쿼리 제거

* feat: 중간에 구간이 추가될 때 거리도 다시 계산해서 넣어주는 로직 추가

* refactor: 가독성 위해 함수실행 flow 변경

* feat: 자하철 역을 기준으로 상행역과 하행역이 구간에 등록되어 있는지 확인하는 쿼리 추가

* feat: 지하철역이 구간에 등록되어 있으면 지하철역을 삭제할 수 없도록 예외 추가

* feat: 지하철 구간 조회 구간순서로 정렬해서 보여주는 로직 추가

* refactor: 안쓰는 매개변수 제거

* refactor: 함수 명 의도에 맞게 수정

* refactor: getter 사용 지양하도록 수정

* refactor: addBranch 함수 실행 분기처리 위한 sections 길이 검증 로직 추가

* feat: 빈 구간 값 정의

* refactor: 갈래길 방지 로직 SectionService -> Sections에 위임

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -3,6 +3,7 @@ package wooteco.subway.service;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
+import wooteco.subway.dao.JdbcSectionDao;
 import wooteco.subway.dao.JdbcStationDao;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
@@ -13,9 +14,11 @@ public class StationService {
 
     private static final String STATION_DUPLICATION = "이미 등록된 지하철 역입니다.";
     private final JdbcStationDao stationDao;
+    private final JdbcSectionDao sectionDao;
 
-    public StationService(JdbcStationDao stationDao) {
+    public StationService(JdbcStationDao stationDao, JdbcSectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     public StationResponse createStation(StationRequest stationRequest) {
@@ -31,6 +34,11 @@ public class StationService {
         }
     }
 
+    public StationResponse getStation(Long id) {
+        Station station = stationDao.findById(id);
+        return new StationResponse(station.getId(), station.getName());
+    }
+
     public List<StationResponse> getStations() {
         List<Station> stations = stationDao.findAll();
         List<StationResponse> stationResponses = stations.stream()
@@ -40,6 +48,9 @@ public class StationService {
     }
 
     public boolean deleteStation(Long id) {
+        if (sectionDao.isExistByStationId(id)) {
+            throw new IllegalArgumentException("구간이 등록되어 있기 때문에 역을 삭제할 수 없습니다.");
+        }
         return stationDao.deleteById(id);
     }
 }
